+++ src/lib/algorithms/sequential/pushrelabel.h
@@ -7,18 +7,18 @@ class PushRelabelSequentialSolver{
   public: 
     void pushRelabel(MaxFlowInstance *input, MaxFlowSolution *output); 
   private:
-    float** flows; // x coordinate is the source, y coordinate is the sink, for edges 
-    float *excessPerVertex; // excess flow on each vertex also exists 
+    int** flows; // x coordinate is the source, y coordinate is the sink, for edges 
+    int *excessPerVertex; // excess flow on each vertex also exists 
     int *d; // the labels 
     int *active; // replace with a queue 
     queue<int> activeQueue; 
-    float totalFlow; 
+    int totalFlow; 
     Timer t; 
 
     void initialize(MaxFlowInstance *input);
     void preflow(MaxFlowInstance *input);
-    bool push(int numVertices, float **cap, int u, int sink);
-    void relabel(int numVertices, float **cap, int u);
+    bool push(int numVertices, int **cap, int u, int sink);
+    void relabel(int numVertices, int **cap, int u);
     int existsActiveNode(MaxFlowInstance *input);
 };
